// src/api/generated/models/signalDevJobVacancyInterviewRequestType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Brew
 * OpenAPI spec version: 0.0.1
 */

/**
 * Tipo do sinal WebRTC repassado para o outro participante
 */
export type SignalDevJobVacancyInterviewRequestType = typeof SignalDevJobVacancyInterviewRequestType[keyof typeof SignalDevJobVacancyInterviewRequestType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SignalDevJobVacancyInterviewRequestType = {
  offer: 'offer',
  answer: 'answer',
  candidate: 'candidate',
} as const;